audioNode/requirements.txt
altgraph==0.17
click==8.0.1
Flask==2.0.1
future==0.18.2
gevent==21.1.2
greenlet==1.1.0
grpcio==1.38.1
grpcio-tools==1.38.0
itsdangerous==2.0.1
Jinja2==3.0.1
macholib==1.14
MarkupSafe==2.0.1
msgpack==1.0.2
protobuf==3.17.3
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyzmq==22.1.0
six==1.16.0
Werkzeug==2.0.1
zerorpc==0.6.3
zope.event==4.5.0
zope.interface==5.4.0
sounddevice==0.3.12
soundFile===0.10.2
